# campus-lost-found/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Lost & Found — API
  version: 1.0.0
  description: Backend API for Section A — Lost & Found (Auth, Items, Claims)

servers:
  - url: http://127.0.0.1:3000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string

    TokenResponse:
      type: object
      properties:
        token:
          type: string

    Role:
      type: string
      nullable: true

    ItemType:
      type: string
      enum: [FOUND, LOST]

    ClaimStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]

    UserLite:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        name: { type: string, nullable: true }

    ItemCreateRequest:
      type: object
      required: [type, title]
      properties:
        type: { $ref: '#/components/schemas/ItemType' }
        title: { type: string }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true }
        locationFound: { type: string, nullable: true }
        dateFound: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }

paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200': { description: OK }
  /api/auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
                name: { type: string }
                gender: { type: string, nullable: true }
                registerNumber: { type: string, nullable: true }
                faculty: { type: string, nullable: true }
                phone: { type: string, nullable: true }
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '201': { description: Created }
        '409':
          description: User exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/auth/login:
    post:
      summary: Login (returns JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/items:
    get:
      summary: List items (public)
      parameters:
        - in: query
          name: type
          schema: { $ref: '#/components/schemas/ItemType' }
          required: false
      responses:
        '200': { description: OK }

    post:
      summary: Create item (auth)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ItemCreateRequest' }
      responses:
        '201': { description: Created }
        '401': { description: Unauthorized }

  /api/items/{id}:
    get:
      summary: Get item details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Not found }
  /api/items/{id}/claim:
    post:
      summary: Create claim
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                message: { type: string }
                contact: { type: string }
      responses:
        '201': { description: Created }
        '401': { description: Unauthorized }

  /api/items/{id}/claims:
    get:
      summary: List claims for an item (owner only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }

  /api/claims/mine:
    get:
      summary: My claims
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }

  /api/claims/{id}/approve:
    post:
      summary: Approve claim
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }

  /api/claims/{id}/reject:
    post:
      summary: Reject claim
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }
  /api/upload:
    post:
      summary: Upload image file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Uploaded OK
        '400': { description: Bad request }